warnings/compilation fixed

+++ Examples/Algorithms/TrackFinding/src/SeedingAlgorithm.cpp
@@ -94,7 +94,7 @@ ActsExamples::SeedingAlgorithm::SeedingAlgorithm(
 //  }
 
   if (m_cfg.seedFinderConfig.zBinsCustomLooping.size() != 0) {
-		for (int i=1; i!=m_cfg.gridConfig.zBinEdges.size(); i++) {
+		for (size_t i=1; i!=m_cfg.gridConfig.zBinEdges.size(); i++) {
 			if(std::find(m_cfg.seedFinderConfig.zBinsCustomLooping.begin(), m_cfg.seedFinderConfig.zBinsCustomLooping.end(), i) == m_cfg.seedFinderConfig.zBinsCustomLooping.end()) {
 				throw std::invalid_argument("Inconsistent config zBinsCustomLooping");
 			}
